Add files via upload
This is my modification of the lohboot function according to the article by Loh in 2010. The former function did not consider a block bootsrap but resampled the local contributions of each point which does not allow to capture the dependence structure of the process. 
Now, the simulation in Loh's paper are reproductible with the lohboot function.